% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_duck_r.R
\name{attach_db}
\alias{attach_db}
\title{Attach a database to the currently running DuckDB and update the search_path
to make the objects in the new database available without using fully
qualified names}
\usage{
attach_db(
  conn_obj,
  duckdb_path,
  duckdb_alias = make.names(fs::path_ext_remove(fs::path_file(duckdb_path))),
  read_only = TRUE
)
}
\arguments{
\item{conn_obj}{DBI connection to DuckDB}

\item{duckdb_path}{path to the DuckDB file to attach}

\item{duckdb_alias}{alias to use fir the newly attached database}

\item{read_only}{whether it should be attached in read only mode}
}
\description{
Attach a database to the currently running DuckDB and update the search_path
to make the objects in the new database available without using fully
qualified names
}
