% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookme.R
\name{bookme}
\alias{bookme}
\title{Wrapper of bookdown::render_book, running it in input_dir}
\usage{
bookme(input_dir, input_files = "*", output_dir = NULL, output_format = NULL)
}
\arguments{
\item{input_dir}{the main directory of the book}

\item{input_files}{character vector with input files, in case you do not
want to render them all (the default)}

\item{output_dir}{character vector where to  put the rendered book}

\item{output_format}{https://bookdown.org/yihui/bookdown/output-formats.html}
}
\value{
character vector with the path to the output
}
\description{
bookdown::render_book renders "multiple R Markdown files
under the current working directory into a book".
TODO: it seems this restriction will be lifted in the next bookdown release
so let's see when it is out and adjust here as necessary
That is, IMHO, unfortunate. In particular, when the book
is not the root of your project, you would have to mess with
setwd() or the like and {here} won't help you either.
This function works around this issue by leveraging xfun::in_dir
to run bookdown::render_book in the input_dir of your choice.
Then I do not have to deal myself with setwd() and changing it back.
}
\details{
In addition, the function calls base::shell to open the book in the
default viewer for the output
}
\examples{
bookme("report")
}
