% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_duck_r.R
\name{duck_tar_format}
\alias{duck_tar_format}
\title{A custom format. The write function is standard, as we only want to store
in the target the info that allows us to create the connection. The read
function would do the heavy-lifting.}
\usage{
duck_tar_format
}
\description{
A custom format. The write function is standard, as we only want to store
in the target the info that allows us to create the connection. The read
function would do the heavy-lifting.
}
