% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowme-package.R
\docType{package}
\name{flowme-package}
\alias{flowme}
\alias{flowme-package}
\title{flowme: Add boilerplate code to your project for some specific workflows}
\description{
One function call to include in your current project tailored files and directories to support some specific workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://edalfon.github.io/flowme/}
  \item \url{https://github.com/edalfon/flowme}
  \item Report bugs at \url{https://github.com/edalfon/flowme/issues}
}

}
\author{
\strong{Maintainer}: Eduardo Alfonso-Sierra \email{edalfon@gmail.com} (\href{https://orcid.org/0000-0001-6430-5135}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Eduardo Alfonso-Sierra \email{edalfon@gmail.com} (\href{https://orcid.org/0000-0001-6430-5135}{ORCID}) [copyright holder]
}

}
\keyword{internal}
