% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_duck_r.R
\name{get_search_path}
\alias{get_search_path}
\title{Query  DuckDB's \code{search_path}}
\usage{
get_search_path(conn_obj)
}
\arguments{
\item{conn_obj}{DBI connection to DuckDB}
}
\description{
Query  DuckDB's \code{search_path}
}
