% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{ignore_drake}
\alias{ignore_drake}
\title{Add entries to .gitignore, to ignore non-version-control-friendly files
in this drake-based workflow}
\usage{
ignore_drake()
}
\description{
Add entries to .gitignore, to ignore non-version-control-friendly files
in this drake-based workflow
}
