% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_duck_r.R
\name{set_search_path}
\alias{set_search_path}
\title{Set DuckDB's \code{search_path}}
\usage{
set_search_path(conn_obj, new_search_path)
}
\arguments{
\item{conn_obj}{DBI connection to DuckDB}

\item{new_search_path}{character with the new search path (comma-separated
list of catalogs)}
}
\description{
Set DuckDB's \code{search_path}
}
