% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_duck_r.R
\name{update_search_path}
\alias{update_search_path}
\title{Update DuckDB's \code{search_path} by prepending and/or appending values}
\usage{
update_search_path(conn_obj, prepend = NULL, append = NULL)
}
\arguments{
\item{conn_obj}{DBI connection to DuckDB}

\item{prepend}{character with value to prepend}

\item{append}{character with value to append}
}
\description{
Update DuckDB's \code{search_path} by prepending and/or appending values
}
