% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{use_drake_description}
\alias{use_drake_description}
\title{Create a description file, including key dependencies to run a drake workflow}
\usage{
use_drake_description(install_deps = TRUE)
}
\arguments{
\item{install_deps}{logical, whether to install drake and other dependencies}
}
\description{
TODO: restart rstudio, if available? should we?
}
