% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{use_drake_gitignore}
\alias{use_drake_gitignore}
\title{Add entries to .gitignore, to ignore non-version-control-friendly files
in this drake-based workflow}
\usage{
use_drake_gitignore()
}
\description{
Add entries to .gitignore, to ignore non-version-control-friendly files
in this drake-based workflow
}
