% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drake.R
\name{use_drake_templates}
\alias{use_drake_templates}
\title{Copy drake templates to the current project}
\usage{
use_drake_templates()
}
\description{
Leverage usethis::use_template to use flowme's drake templates
TODO: let the user customize the "report" folder
TODO: more convenient overwrite?
}
