% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{use_targets_description}
\alias{use_targets_description}
\title{Create a description file, including key dependencies to run the workflow}
\usage{
use_targets_description(install_deps = TRUE)
}
\arguments{
\item{install_deps}{logical, whether to install dependencies}
}
\description{
TODO: restart rstudio, if available? should we?
}
