% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{use_targets_gitignore}
\alias{use_targets_gitignore}
\title{Add entries to .gitignore, to ignore non-version-control-friendly files
in this targets-based workflow}
\usage{
use_targets_gitignore()
}
\description{
Add entries to .gitignore, to ignore non-version-control-friendly files
in this targets-based workflow
}
