% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{use_targets_templates}
\alias{use_targets_templates}
\title{Copy \code{targets} templates to the current project}
\usage{
use_targets_templates()
}
\description{
TODO: let the user customize the "report" folder
TODO: more convenient overwrite?
}
