% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowme-package.R
\docType{package}
\name{flowme-package}
\alias{flowme}
\alias{flowme-package}
\title{flowme: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Eduardo Alfonso-Sierra \email{edalfon@gmail.com} (\href{https://orcid.org/0000-0001-6430-5135}{ORCID}) [copyright holder]

}
\keyword{internal}
